+++ packages/cdk/lib/construct/auth.ts
@@ -1,0 +1,167 @@
+import {
+  StringAttribute,
+  UserPool,
+  UserPoolClient,
+} from 'aws-cdk-lib/aws-cognito';
+import {
+  IdentityPool,
+  UserPoolAuthenticationProvider,
+} from 'aws-cdk-lib/aws-cognito-identitypool';
+import { Construct } from 'constructs';
+import { Duration } from 'aws-cdk-lib';
+import {
+  CfnRole,
+  Effect,
+  Policy,
+  PolicyStatement,
+  Role,
+} from 'aws-cdk-lib/aws-iam';
+
+type AuthProps = {
+  samlAuthEnabled: boolean;
+  samlDefaultAuthEnabled: boolean;
+  selfSignUpEnabled: boolean;
+
+  allowedIpV4AddressRanges?: string[] | null;
+  allowedIpV6AddressRanges?: string[] | null;
+};
+
+class Auth extends Construct {
+  readonly userPool: UserPool;
+  readonly client: UserPoolClient;
+  readonly idPool: IdentityPool;
+
+  constructor(scope: Construct, id: string, props: AuthProps) {
+    super(scope, id);
+
+    const {
+      samlAuthEnabled,
+      samlDefaultAuthEnabled,
+      selfSignUpEnabled,
+      allowedIpV4AddressRanges,
+      allowedIpV6AddressRanges,
+    } = props;
+
+    const userPool = new UserPool(this, 'UserPool', {
+      // If SAML authentication is enabled and default auth is disabled, do not use self-sign-up with UserPool. Be aware of security.
+      selfSignUpEnabled:
+        samlAuthEnabled && !samlDefaultAuthEnabled ? false : selfSignUpEnabled,
+      signInAliases: {
+        username: false,
+        email: true,
+      },
+      passwordPolicy: {
+        requireUppercase: true,
+        requireSymbols: true,
+        requireDigits: true,
+        minLength: 8,
+      },
+      customAttributes: {
+        tenant_id: new StringAttribute({
+          minLen: 1,
+          maxLen: 50,
+          mutable: true,
+        }),
+        tenantAdmin: new StringAttribute({
+          minLen: 4, // "true" or "false"
+          maxLen: 5,
+          mutable: true, // Allows updating admin status
+        }),
+      },
+    });
+
+    const client = userPool.addClient('client', {
+      idTokenValidity: Duration.days(1),
+      refreshTokenValidity: Duration.days(30),
+      accessTokenValidity: Duration.hours(1),
+      enableTokenRevocation: true,
+      authFlows: {
+        adminUserPassword: true,
+        custom: true,
+        userPassword: true,
+        userSrp: true,
+      },
+    });
+
+    const idPool = new IdentityPool(this, 'IdentityPool', {
+      authenticationProviders: {
+        userPools: [
+          new UserPoolAuthenticationProvider({
+            userPool,
+            userPoolClient: client,
+          }),
+        ],
+      },
+      allowUnauthenticatedIdentities: false,
+    });
+
+    // Fix the trust relationship for the authenticated role
+    // The Identity Pool's default authenticated role needs proper trust policy
+    const authenticatedRole = idPool.authenticatedRole as Role;
+    const cfnRole = authenticatedRole.node.defaultChild as CfnRole;
+
+    // Update the assume role policy to properly trust cognito-identity.amazonaws.com
+    cfnRole.assumeRolePolicyDocument = {
+      Version: '2012-10-17',
+      Statement: [
+        {
+          Effect: 'Allow',
+          Principal: {
+            Federated: 'cognito-identity.amazonaws.com',
+          },
+          Action: ['sts:AssumeRoleWithWebIdentity', 'sts:TagSession'],
+          Condition: {
+            StringEquals: {
+              'cognito-identity.amazonaws.com:aud': idPool.identityPoolId,
+            },
+            'ForAnyValue:StringLike': {
+              'cognito-identity.amazonaws.com:amr': 'authenticated',
+            },
+          },
+        },
+      ],
+    };
+
+    if (allowedIpV4AddressRanges || allowedIpV6AddressRanges) {
+      const ipRanges = [
+        ...(allowedIpV4AddressRanges ? allowedIpV4AddressRanges : []),
+        ...(allowedIpV6AddressRanges ? allowedIpV6AddressRanges : []),
+      ];
+
+      idPool.authenticatedRole.attachInlinePolicy(
+        new Policy(this, 'SourceIpPolicy', {
+          statements: [
+            new PolicyStatement({
+              effect: Effect.DENY,
+              resources: ['*'],
+              actions: ['*'],
+              conditions: {
+                NotIpAddress: {
+                  'aws:SourceIp': ipRanges,
+                },
+              },
+            }),
+          ],
+        })
+      );
+    }
+
+    idPool.authenticatedRole.attachInlinePolicy(
+      new Policy(this, 'PollyPolicy', {
+        statements: [
+          new PolicyStatement({
+            effect: Effect.ALLOW,
+            resources: ['*'],
+            actions: ['polly:SynthesizeSpeech'],
+          }),
+        ],
+      })
+    );
+
+    this.client = client;
+    this.userPool = userPool;
+    this.idPool = idPool;
+  }
+}
+
+export default Auth;
