+++ packages/cdk/lib/stacks/common/auth-stack.ts
@@ -1,27 +1,9 @@
-import { CfnOutput, Duration, Stack, StackProps } from 'aws-cdk-lib';
-import {
-  LambdaVersion,
-  StringAttribute,
-  UserPool,
-  UserPoolClient,
-  UserPoolOperation,
-} from 'aws-cdk-lib/aws-cognito';
-import {
-  IdentityPool,
-  UserPoolAuthenticationProvider,
-} from 'aws-cdk-lib/aws-cognito-identitypool';
-import {
-  Effect,
-  Policy,
-  PolicyStatement,
-  Role,
-  CfnRole,
-} from 'aws-cdk-lib/aws-iam';
+import { CfnOutput, Stack, StackProps } from 'aws-cdk-lib';
+import { UserPool, UserPoolClient } from 'aws-cdk-lib/aws-cognito';
+import { IdentityPool } from 'aws-cdk-lib/aws-cognito-identitypool';
 import { Construct } from 'constructs';
-import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
-import { LAMBDA_RUNTIME_NODEJS, LAMBDA_RUNTIME_PYTHON } from '../../../consts';
-import { PythonFunction } from '@aws-cdk/aws-lambda-python-alpha';
 import { ProcessedStackInput } from '../../stack-input';
+import Auth from '../../construct/auth';
 
 interface AuthStackProps extends StackProps {
   readonly params: ProcessedStackInput;
@@ -37,179 +19,17 @@ class AuthStack extends Stack {
 
     const { params } = props;
 
-    const userPool = new UserPool(this, 'UserPool', {
-      // If SAML authentication is enabled and default auth is disabled, do not use self-sign-up with UserPool. Be aware of security.
-      selfSignUpEnabled:
-        params.samlAuthEnabled && !params.samlDefaultAuthEnabled
-          ? false
-          : params.selfSignUpEnabled,
-      signInAliases: {
-        username: false,
-        email: true,
-      },
-      passwordPolicy: {
-        requireUppercase: true,
-        requireSymbols: true,
-        requireDigits: true,
-        minLength: 8,
-      },
-      customAttributes: {
-        tenant_id: new StringAttribute({
-          minLen: 1,
-          maxLen: 50,
-          mutable: true,
-        }),
-        tenantAdmin: new StringAttribute({
-          minLen: 4, // "true" or "false"
-          maxLen: 5,
-          mutable: true, // Allows updating admin status
-        }),
-      },
+    const auth = new Auth(this, 'Auth', {
+      samlAuthEnabled: params.samlAuthEnabled,
+      samlDefaultAuthEnabled: params.samlDefaultAuthEnabled,
+      selfSignUpEnabled: params.selfSignUpEnabled,
+      allowedIpV4AddressRanges: params.allowedIpV4AddressRanges,
+      allowedIpV6AddressRanges: params.allowedIpV6AddressRanges,
     });
 
-    const client = userPool.addClient('client', {
-      idTokenValidity: Duration.days(1),
-      refreshTokenValidity: Duration.days(30),
-      accessTokenValidity: Duration.hours(1),
-      enableTokenRevocation: true,
-      authFlows: {
-        adminUserPassword: true,
-        custom: true,
-        userPassword: true,
-        userSrp: true,
-      },
-    });
-
-    const idPool = new IdentityPool(this, 'IdentityPool', {
-      authenticationProviders: {
-        userPools: [
-          new UserPoolAuthenticationProvider({
-            userPool,
-            userPoolClient: client,
-          }),
-        ],
-      },
-      allowUnauthenticatedIdentities: false,
-    });
-
-    // Fix the trust relationship for the authenticated role
-    // The Identity Pool's default authenticated role needs proper trust policy
-    const authenticatedRole = idPool.authenticatedRole as Role;
-    const cfnRole = authenticatedRole.node.defaultChild as CfnRole;
-
-    // Update the assume role policy to properly trust cognito-identity.amazonaws.com
-    cfnRole.assumeRolePolicyDocument = {
-      Version: '2012-10-17',
-      Statement: [
-        {
-          Effect: 'Allow',
-          Principal: {
-            Federated: 'cognito-identity.amazonaws.com',
-          },
-          Action: ['sts:AssumeRoleWithWebIdentity', 'sts:TagSession'],
-          Condition: {
-            StringEquals: {
-              'cognito-identity.amazonaws.com:aud': idPool.identityPoolId,
-            },
-            'ForAnyValue:StringLike': {
-              'cognito-identity.amazonaws.com:amr': 'authenticated',
-            },
-          },
-        },
-      ],
-    };
-
-    if (params.allowedIpV4AddressRanges || params.allowedIpV6AddressRanges) {
-      const ipRanges = [
-        ...(params.allowedIpV4AddressRanges
-          ? params.allowedIpV4AddressRanges
-          : []),
-        ...(params.allowedIpV6AddressRanges
-          ? params.allowedIpV6AddressRanges
-          : []),
-      ];
-
-      idPool.authenticatedRole.attachInlinePolicy(
-        new Policy(this, 'SourceIpPolicy', {
-          statements: [
-            new PolicyStatement({
-              effect: Effect.DENY,
-              resources: ['*'],
-              actions: ['*'],
-              conditions: {
-                NotIpAddress: {
-                  'aws:SourceIp': ipRanges,
-                },
-              },
-            }),
-          ],
-        })
-      );
-    }
-
-    idPool.authenticatedRole.attachInlinePolicy(
-      new Policy(this, 'PollyPolicy', {
-        statements: [
-          new PolicyStatement({
-            effect: Effect.ALLOW,
-            resources: ['*'],
-            actions: ['polly:SynthesizeSpeech'],
-          }),
-        ],
-      })
-    );
-
-    // Lambda
-    if (params.selfSignUpTenantMap && params.selfSignUpTenantMap.length > 0) {
-      const checkTenantFunction = new NodejsFunction(this, 'CheckTenant', {
-        runtime: LAMBDA_RUNTIME_NODEJS,
-        entry: './lambda/checkTenant.ts',
-        timeout: Duration.seconds(30),
-        environment: {
-          SELF_SIGNUP_TENANT_MAP: JSON.stringify(params.selfSignUpTenantMap),
-        },
-      });
-
-      userPool.addTrigger(UserPoolOperation.PRE_SIGN_UP, checkTenantFunction);
-
-      const assignTenantFunction = new NodejsFunction(this, 'AssignTenant', {
-        runtime: LAMBDA_RUNTIME_NODEJS,
-        entry: './lambda/assignTenant.ts',
-        timeout: Duration.seconds(30),
-        environment: {
-          SELF_SIGNUP_TENANT_MAP: JSON.stringify(params.selfSignUpTenantMap),
-        },
-      });
-
-      assignTenantFunction.addToRolePolicy(
-        new PolicyStatement({
-          effect: Effect.ALLOW,
-          actions: ['cognito-idp:AdminUpdateUserAttributes'],
-          resources: ['*'],
-        })
-      );
-      userPool.addTrigger(
-        UserPoolOperation.POST_CONFIRMATION,
-        assignTenantFunction
-      );
-    }
-
-    // Pre Token Generation Lambda for adding custom claims
-    const preTokenGenerationFunction = new PythonFunction(
-      this,
-      'PreTokenGeneration',
-      {
-        runtime: LAMBDA_RUNTIME_PYTHON,
-        entry: './lambda/pre_token_generation',
-        timeout: Duration.seconds(5),
-      }
-    );
-
-    userPool.addTrigger(
-      UserPoolOperation.PRE_TOKEN_GENERATION_CONFIG,
-      preTokenGenerationFunction,
-      LambdaVersion.V2_0
-    );
+    const userPool = auth.userPool;
+    const client = auth.client;
+    const idPool = auth.idPool;
 
     new CfnOutput(this, 'UserPoolId', { value: userPool.userPoolId });
